Added explosions

+++ Common/IImage.cs
@@ -15,6 +15,6 @@
         void Draw(byte[] buffer);
         IImage Scale(float scale);
         IImage Rotate(float angle);
-        IImage Fade(byte fade);
+        IImage Fade(float fade);
     }
 }
